--- app/update/delete_handler.py
@@ -1,4 +1,0 @@
-
-
-def lambda_handler(event, context):
-    return success_response(200, "Product deleted successfully")
