--- lambda-layers/flask-layer/requirements.txt
@@ -1,10 +1,0 @@
-Flask==3.0.3
-Werkzeug==3.0.6
-aws-wsgi==0.2.7
-aws-lambda-powertools==2.30.1
-aws-xray-sdk==2.12.1
-python-dotenv==1.0.0
-PyJWT==2.8.0
-Flask==3.0.3
-blinker==1.7.0
-typing-extensions==4.9.0
