+++ src/delete/delete_handler.py
