+++ src/get/get_all/requirements.txt
@@ -1,0 +1,1 @@
+aws-lambda-powertools
