+++ src/get/get_by_pk/requirements.txt
@@ -1,0 +1,1 @@
+aws-lambda-powertools
